% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funredun.R
\name{funredun}
\alias{funredun}
\title{Functional Redundancy}
\usage{
funredun(spDat, funDat)
}
\arguments{
\item{spDat}{Data frame with rows as sites, columns as species, and elements as counts}

\item{funDat}{Data frame with rows as species (same as spDat column names), columns as traits, elements as counts, measures, binary, etc.}
}
\value{
A data frame with rows as sites and a column of functional redundancy
}
\description{
This function calculates functional redundancy as Simpson's D minus
dissimilarity between species weighted by the proportions of the species.
}
